+++ eras/babbage/impl/CHANGELOG.md
@@ -29,6 +29,10 @@
 
 ### `testlib`
 
+* Add  to `Test.Cardano.Ledger.Babbage.Examples`:
+  - `exampleBabbagePParamsUpdate`
+  - `exampleBabbageOnwardsEraPParams`
+  - `exampleBabbageOnwardsEraPParamsUpdate`
 * Add `proxy era` argument to `babbageEraSpecificSpec`
 * Make `Test.Cardano.Ledger.Babbage.Imp.spec` accept `Proxy era`
 * Add `Test.Cardano.Ledger.Babbage.Imp.PoolSpec` module with `babbageEraSpecificSpec`